refactor(app): split the event flow out of KiwiCore.swift
KiwiCore.swift sat one line under the 350 hard ceiling (§2.1)
after #143's additions; the wiring file is not the cohesive
performance-critical case the ceiling band is reserved for.
handle(_:) and restore(_:) move to KiwiCore+Events.swift
unchanged (handle drops `private` for cross-file extension
visibility); KiwiCore.swift lands back at 217 lines.

refs #143
