+++ Sources/KiwiDeskCore/App/KiwiCore.swift
@@ -152,8 +152,7 @@ public final class KiwiCore {
             self?.moveBarItem(space: space, from: from, to: to)
         }
         tiler.animation.onAllAnimationsEnded = { [weak self] in
-            self?.runPendingZOrderRestore()
-            self?.runPendingFocusRaise()
+            self?.animationsDidSettle()
         }
 
         socket.handler = { [weak self] command, args in
@@ -201,138 +200,6 @@ public final class KiwiCore {
         // for them.
         persistScrollOffset()
         updateAppBar()
-    }
-
-    // MARK: - Event flow
-
-    private func handle(_ event: KiwiEvent) {
-        // Closing or minimizing the focused window must hand
-        // focus to the space's fallback window (state already
-        // picks one; the AX raise below makes it real). The
-        // gone-event payload needs the window's app and space
-        // captured here too — state.apply removes both.
-        var focusLost = false
-        var goneApp: String? = nil
-        var goneSpace: SpaceID? = nil
-        if case .windowDestroyed(let id, _) = event {
-            focusLost = activeSpace?.focused == id
-            goneApp = state.windows[id]?.appName
-            goneSpace = state.workspaces.space(of: id)
-        }
-        state.spawnPlacements = [
-            .bsp: tiler.settings.bsp.newWindowPlacement,
-            .stack: tiler.settings.stack.newWindowPlacement,
-            .scrolling:
-                tiler.settings.scrolling.newWindowPlacement,
-            .grid: tiler.settings.grid.newWindowPlacement,
-        ]
-        state.spawnOverride = tiler.settings.placementOverride
-        state.apply(event)
-        var newlyCreatedWindow: WindowID? = nil
-        switch event {
-        case .displaysChanged:
-            tiler.displaysChanged()
-            handleMonitorChange()
-            emitMonitorChange()
-        case .windowFocused(let id):
-            // A real focus echo supersedes a pending deferred
-            // raise (#143): the OS already raised whatever the
-            // user reached (a click mid-pan), and a stale raise
-            // firing after the pan would steal focus back.
-            pendingFocusRaise = nil
-            emitFocusChange(id)
-            // cmd+tab (or a click) can reach a window hidden
-            // in an inactive virtual space; pull that space
-            // forward instead of typing into a stashed window.
-            // Deferred: app activation transiently re-reports
-            // the app's OLD focused window right before a new
-            // window opens — only follow if focus settles.
-            if let space = state.workspaces.space(of: id),
-                space != state.workspaces.activeSpace
-            {
-                scheduleFocusFollow(id)
-            } else if activeSpace?.mode.isFocusDriven == true {
-                retileWithScrollDuration()
-            }
-        case .windowCreated(let window):
-            // A brand-new window supersedes a pending follow
-            // of a hidden window (see above).
-            deferred.cancel(.focusFollow)
-            newlyCreatedWindow = window.id
-            emitWindowCreated(window)
-        case .windowMoved(let id, let frame):
-            drag.windowMoved(id, frame: frame)
-        case .windowResized(let id, let frame):
-            // Resize gestures share the drag pipeline (same
-            // settle debounce). Only mouse-driven resizes
-            // count; apps resizing themselves are corrected
-            // by the next retile. `validated` lets the
-            // trailing events of a fast resize (classified
-            // via the recent press near a slot edge) start
-            // the gesture even after the release.
-            if isResizeGesture(id) {
-                drag.windowMoved(
-                    id,
-                    frame: frame,
-                    validated: true
-                )
-            }
-        case .windowDestroyed(let id, let wasMinimized):
-            drag.cancel(id)
-            dragOverlay.hideAll()
-            if wasMinimized {
-                emitWindowMinimized(
-                    id,
-                    app: goneApp,
-                    space: goneSpace
-                )
-            } else {
-                emitWindowDestroyed(
-                    id,
-                    app: goneApp,
-                    space: goneSpace
-                )
-            }
-        case .nativeSpaceChanged:
-            handleNativeSpaceChange()
-        default:
-            break
-        }
-        if TilingEngine.shouldRetile(after: event) {
-            retile(newlyCreatedWindow: newlyCreatedWindow)
-        }
-        // Only raise windows the app still lists: after a
-        // native Space switch the fallback may live on the
-        // previous desktop, and raising it would switch back.
-        if focusLost, let next = activeSpace?.focused,
-            eventLoop.isListed(next)
-        {
-            focusWindow(next)
-        }
-    }
-
-    /// Re-applies a snapshot after wake/unlock, a crash, or a
-    /// restart: space membership and order first (the array
-    /// order is the layout order), then the raw frames. Frames
-    /// go through the tiler's frame pipeline so the resulting
-    /// AX echoes are not mistaken for user drags.
-    func restore(_ snapshot: StateSnapshot) {
-        state.adopt(snapshot)
-        for record in snapshot.windows {
-            tiler.setFrame(record.windowID, record.frame)
-        }
-        // Diagnostic: snapshot windows that are not tracked
-        // yet stay out of their space until a reconcile finds
-        // them (cold-AX startup scan, issue #21 follow-up).
-        let missing = snapshot.windows.filter {
-            state.windows[$0.windowID] == nil
-        }.count
-        if missing > 0 {
-            onLog(
-                "restore: \(missing) snapshot windows not "
-                    + "tracked yet"
-            )
-        }
     }
 
     // MARK: - Accessors
